Lists Advanced - Lab
Conditions:
You will be receiving to-do notes until you get the command "End". The notes will be in the format: "{importance}-{note}". 
Return a list containing all to-do notes sorted by importance in ascending order. 
The importance value will always be an integer between 1 and 10 (inclusive).
Hint:
Use the pop() and insert() methods.