Problems for Champions Part II
Problem:
We all know the game called "Bulls and Cows" (https://en.wikipedia.org/wiki/Bulls_and_cows). Upon having a particular 4-digit secret number and a 4-digit suggested number, the following rules are applied:
- If a digit in the suggested number matches a digit in the secret number and is located at the same position, we have a bull.
- If a digit in the suggested number matches a digit in the secret number but is located at a different position, we have a cow.

Upon having a particular secret number and the bulls and cows pertaining to it, our task is to find all possible suggested numbers in ascending order. If there are no suggested numbers that match the criteria provided from the console, we must print "No".

Input Data:
The input data is read from the console. We have 3 lines in the input data:
- The first contains the secret number.
- The second contains the number of bulls.
- The third contains the number of cows.
- The input data will always be valid. There is no need to verify them.

Output Data:
The output data must be printed on the console. The output must consist of a single line - all suggested numbers, space-separated. If there are no suggested numbers that match the criteria provided from the console, we must print “No”.

Constraints:
- The secret number will always consist of 4 digits in the range [1..9].
- The number of cows and bulls will always be in the range [0..9].
- Allowed execution time: 0.15 seconds.
- Allowed memory: 16 MB.